Handle keyboard events.
Note that this was rather more difficult than it needed to
be because I didn't realise that the gemlib binding for
evnt_multi has a different signature from that in the GEM
docs; Interval is of type unsigned long, rather than being
composed of two parameters, a high word and a low word.
C doesn't give you an error message when you screw up
stuff like this. :-(