Get rid of those pesky multiply operations
Instead, we just keep incrementing the index every time we move
from cell to cell.

And if a cell is equal to zero (which most of them are) then we
immediately skip it.

A further optimisation (not yet implemented) would be to skip
four cells at time if a whole longword is equal to zero. Might
have to stop using the index and start using an actual raw
pointer if we want to do that. :-o