Support larger grid sizes
Turns out a 16-bit integer isn't enough to index a larger grid.
Thankfully, switching to a 32-bit integer doesn't seem to have screwed
performance on the Falcon because (I think) it just lives in a register.

Should probably do away with indices altogether and just use a pointer,
since that's what it looks like it gets compiled down to anyway.