Optimise the step function just a little
We look at the next four bytes and don't do any work if all are
zero. Shame 68000 registers are only 32-bit because that's all
that we can test in one go.

We could probably test multiple longwords slightly quicker in asm
by using movem to pull the values out of memory into several
registers and then testing each one in turn.

The alternative that I'm seriously considering is using a quad
tree where any completely blank area would be a (hopefully fairly
large) leaf node that we don't need to do much work on.

But for now, it's drawing to the screen that takes most of the
time. So time to optimise that next maybe.