Allocate call_array_copy on the heap
Allocating a huge array on the stack turns out not to be a great idea.
I _think_ that cell_array_copy is so big that it starts overwriting
one of the screen buffers in the BSS. A bit more investigation needed
to confirm that but, in any case, this commit causes the program not
to allocate a huge (up to 307200 byte) array on the stack any more.